refactor(r3f): simplify hello world example
