+++ apps/r3f-hello-world/src/main.tsx
@@ -1,4 +1,6 @@
+import { Canvas } from '@react-three/fiber/webgpu';
 import { StrictMode } from 'react';
+import { Suspense } from 'react';
 import { createRoot } from 'react-dom/client';
 
 import shaperWasmUrl from '@pmndrs/text/text-shaper.wasm?url';
@@ -18,6 +20,16 @@ document.head.append(shaperPreload);
 
 createRoot(root).render(
   <StrictMode>
-    <App />
+    <Canvas
+      camera={{ far: 1_000, near: -1_000, position: [0, 0, 10] }}
+      fallback={<div className="fallback">WebGPU or WebGL2 is required.</div>}
+      flat
+      orthographic
+    >
+      <color attach="background" args={['#07090f']} />
+      <Suspense fallback={null}>
+        <App />
+      </Suspense>
+    </Canvas>
   </StrictMode>,
 );
