+++ docs/packages/r3f-hello-world.md
@@ -5,27 +5,28 @@ description: Demonstrates the public React Three Fiber API with Bitmap, MSDF, Sl
 resource: ../../apps/r3f-hello-world
 workspace_package: '@pmndrs/text-r3f-hello-world'
 documentation_type: reference
-source_digest: 'sha256:d661c8c488210e52695dbcb7d46d106346dc1739a8508434a39ff4d86600b044'
+source_digest: 'sha256:97a2edefbff112367f3ac1675fa267a85d3f2438f2048dc4e2ffb72b808f3977'
 tags: [package, example, react, react-three-fiber, vite]
 sources:
   - id: manifest
     resource: ../../apps/r3f-hello-world/package.json
     title: Example application manifest
   - id: scene
-    resource: ../../apps/r3f-hello-world/src/technique-scene.tsx
+    resource: ../../apps/r3f-hello-world/src/app.tsx
     title: Public R3F technique and fallback example
 generated:
   by: openai-codex/gpt-5.6
-  at: '2026-08-10T03:47:15Z'
+  at: '2026-08-10T16:08:36Z'
 ---
 
 # Package reference: `@pmndrs/text-r3f-hello-world`
 
 This private Vite application is the minimal product-shaped React Three Fiber example. One full-page canvas renders
 `Hello world` through the public `@pmndrs/text/r3f` `Text` component and resolves a Font Awesome globe through an ordered
-font stack. In-canvas MSDF controls replace the rendered text component between Bitmap, MSDF, and Slug; the example does
-not retain a second renderer path or manually pack glyph data. Separate world and UI `TextGroup` roots batch their text
-descendants explicitly; ordinary nested groups only position individual buttons and their background meshes.
+font stack. One `App` component owns the font loads, technique state, one world `Text`, and its in-canvas MSDF controls.
+Changing technique selects the corresponding loaded font stack for that same `Text`; the example does not retain a
+second renderer path or manually pack glyph data. Separate world and UI `TextGroup` roots batch their text descendants
+explicitly; ordinary nested groups only position individual buttons and their background meshes.
 
 The checked-in assets are deliberately bounded at source before baking:
 
@@ -33,8 +34,9 @@ The checked-in assets are deliberately bounded at source before baking:
 - Font Awesome contains six globe and earth PUA scalars, including the displayed `U+F0AC` glyph.
 
 Each GLB embeds Bitmap, MSDF, and Slug raster resources for its subset. The package manifest invokes only the published
-CLI through `pnpm exec text bake`: direct input/output arguments select all three rasters, `--unicodes` delegates shaping-font
-subsetting to pinned HarfBuzz 14.2.0, and `--check` rebakes into temporary storage before requiring byte-identical output.
+CLI through `pnpm exec text bake`: direct input/output arguments select all three rasters, `--unicodes` delegates
+shaping-font subsetting to the package-owned baker Wasm, and `--check` rebakes into temporary storage before requiring
+byte-identical output.
 The example loads each GLB once with one typed raster tuple and receives exact Bitmap, MSDF, and Slug `LoadedFont` values;
 it does not repeat the input URL per technique. Vite emits the public shaper Wasm URL and a combined Inter/Font Awesome
 notice file. Three, React, and React Three Fiber remain ordinary workspace peers rather than part of the core package-size
@@ -43,7 +45,6 @@ graph.
 ## Commands
 
 ```sh
-mise -C apps/benchmarks exec -- node ./scripts/provision-harfbuzz.mts --version=14.2.0
 mise exec -- pnpm --filter @pmndrs/text-r3f-hello-world dev
 mise exec -- pnpm --filter @pmndrs/text-r3f-hello-world check
 ```
